Phase_III_Todo_AI_Chatbot/src/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-multipart==0.0.6
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0
cohere>=5.0.0
openai-agents[litellm]>=0.0.1